style: fix the header
